+++ src/components/ble/MotionService.h
@@ -17,7 +17,7 @@ namespace Pinetime {
       MotionService(Pinetime::System::SystemTask& system, Controllers::MotionController& motionController);
       void Init();
       int OnStepCountRequested(uint16_t connectionHandle, uint16_t attributeHandle, ble_gatt_access_ctxt* context);
-      void OnNewStepCountValue(uint8_t stepCount);
+      void OnNewStepCountValue(uint32_t stepCount);
       void OnNewMotionValues(int16_t x, int16_t y, int16_t z);
 
       void SubscribeNotification(uint16_t connectionHandle, uint16_t attributeHandle);
